Fix the Integration Tests by mocking `socket`
https://www.npmjs.com/package/socket.io-mock
