+++ package.json
@@ -18,6 +18,7 @@
     "prettier": "^2.5.1",
     "sass": "^1.44.0",
     "sass-true": "^6.0.1",
+    "socket.io-mock": "^1.3.2",
     "xunit-viewer": "^7.1.5"
   },
   "scripts": {
